+++ header-head.php
@@ -10,11 +10,8 @@
 #################################################################################################### */
 
 ?><!doctype html>
-<!--[if lt IE 7]> <html class="no-js lt-ie9 lt-ie8 lt-ie7" <?php language_attributes(); ?>> <![endif]-->
-<!--[if IE 7]>    <html class="no-js lt-ie9 lt-ie8" <?php language_attributes(); ?>> <![endif]-->
-<!--[if IE 8]>    <html class="no-js lt-ie9" <?php language_attributes(); ?>> <![endif]-->
-<!--[if gt IE 8]><!--> <html class="no-js" <?php language_attributes(); ?>> <!--<![endif]-->
-<head>
+<html <?php language_attributes(); echo apply_filters( 'tag_html_attr', '' ); ?>>
+<head <?php echo apply_filters( 'tag_head_attr', '' ); ?>>
 
 	<title><?php wp_title(); ?></title>